Migrate code-gen processors to their own project, and separate reflection code.
Add a readme section for including Dagger in your build.

Move most code-gen and reflection handling into separate packages, with polymorphic classes to do the resolution based on provided versions of these classes.

Also move all the logic for determining support for code-gen or not (the specific wiring) into ObjectGraph.get().
