+++ compiler/src/it/missing-at-inject-constructor/pom.xml
@@ -21,15 +21,21 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
   <groupId>com.squareup.dagger.tests</groupId>
-  <artifactId>missing-dependency</artifactId>
   <version>@dagger.version@</version>
   <packaging>jar</packaging>
+  <artifactId>missing-at-inject-constructor</artifactId>
   <name>Dagger Integration Test Basic</name>
   <dependencies>
     <dependency>
       <groupId>com.squareup</groupId>
       <artifactId>dagger</artifactId>
       <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>com.squareup</groupId>
+      <artifactId>dagger-compiler</artifactId>
+      <version>${project.version}</version>
+      <scope>provided</scope>
     </dependency>
   </dependencies>
   <build>
