+++ compiler/src/it/simple-missing-dependency-failure/invoker.properties
