+++ compiler/src/it/simple-missing-dependency-failure/src/main/java/test/TestApp.java
