+++ compiler/src/it/simple-missing-dependency-failure/verify.bsh
